feat: refactored camera configuration
